PP-7077 Reorganise Ledger classes
- Reorganised Ledger related service and models into ledger specific package
