+++ src/main/java/uk/gov/pay/connector/client/ledger/model/TransactionState.java
@@ -1,4 +1,4 @@
-package uk.gov.pay.connector.paritycheck;
+package uk.gov.pay.connector.client.ledger.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
